jQuery y Laravel:Funcionalidad para votar con AJAX

+++ app/Repositories/VoteRepository.php
@@ -22,6 +22,8 @@ class VoteRepository extends BaseRepository
 
   public function unvote(User $user, Ticket $ticket)
   {
+      if (!$user->hasVoted($ticket)) return false;
       $user->voted()->detach($ticket);
+      return true;
   }
 }
